# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Xclusive API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories:
    post:
      operationId: createStory
      tags: [stories]
      summary: Criar story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "201":
          description: Story criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{id}:
    delete:
      operationId: deleteStory
      tags: [stories]
      summary: Eliminar story
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado
  /stories/{id}/view:
    post:
      operationId: viewStory
      tags: [stories]
      summary: Registar visualização de story
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visualização registada
  /stories/{id}/views:
    get:
      operationId: getStoryViews
      tags: [stories]
      summary: Visualizações da story (apenas dono)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de visualizadores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoryView"

  # REELS
  /reels:
    get:
      operationId: getReels
      tags: [reels]
      summary: Feed de reels
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Lista de reels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReels"

    post:
      operationId: createReel
      tags: [reels]
      summary: Publicar reel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReelInput"
      responses:
        "201":
          description: Reel criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reel"
  /reels/{id}:
    get:
      operationId: getReel
      tags: [reels]
      summary: Obter reel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reel"
  /reels/{id}/like:
    post:
      operationId: likeReel
      tags: [reels]
      summary: Curtir reel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Curtido

    delete:
      operationId: unlikeReel
      tags: [reels]
      summary: Remover curtida de reel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Curtida removida

  # EXPLORE
...
  /highlights:
    post:
      operationId: createHighlight
      tags: [highlights]
      summary: Criar destaque de stories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HighlightInput"
      responses:
        "201":
          description: Destaque criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
  /users/{username}/highlights:
    get:
      operationId: getUserHighlights
      tags: [highlights]
      summary: Destaques de um utilizador
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de destaques
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Highlight"
  /highlights/{id}:
    delete:
      operationId: deleteHighlight
      tags: [highlights]
      summary: Eliminar destaque
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado

  # CREATOR / MONETIZATION
...
components:
  schemas:
...
    UserSummary:
      type: object
      required: [id, username, nomeExibicao, verificado]
      properties:
        id:
          type: integer
        username:
          type: string
        nomeExibicao:
          type: string
        avatarUrl:
          type: ["string", "null"]
        verificado:
          type: boolean
        tipoConta:
          type: string
        estaASeguir:
          type: boolean
        totalSeguidores:
          type: integer
...
    Story:
      type: object
      required: [id, autor, mediaUrl, tipo, duracao, audiencia, expirado, visto, criadoEm]
      properties:
        id:
          type: integer
        autor:
          $ref: "#/components/schemas/UserSummary"
        mediaUrl:
          type: string
        tipo:
          type: string
          enum: [imagem, video]
        duracao:
          type: integer
        audiencia:
          type: string
          enum: [todos, subscritores, proximos]
        expirado:
          type: boolean
        visto:
          type: boolean
        totalVisualizacoes:
          type: integer
        expiraEm:
          type: string
          format: date-time
        criadoEm:
          type: string
          format: date-time
...
    StoryInput:
      type: object
      required: [mediaUrl, tipo]
      properties:
        mediaUrl:
          type: string
        tipo:
          type: string
          enum: [imagem, video]
        duracao:
          type: integer
          default: 5
        audiencia:
          type: string
          enum: [todos, subscritores, proximos]
          default: todos
    StoryView:
      type: object
      required: [utilizador, vistoEm]
      properties:
        utilizador:
          $ref: "#/components/schemas/UserSummary"
        vistoEm:
          type: string
          format: date-time

    # REELS
    Reel:
      type: object
      required: [id, autor, videoUrl, exclusivo, totalCurtidas, totalComentarios, curtido, guardado, criadoEm]
      properties:
        id:
          type: integer
        autor:
          $ref: "#/components/schemas/UserSummary"
        videoUrl:
          type: string
        capaUrl:
          type: ["string", "null"]
        legenda:
          type: ["string", "null"]
        somTitulo:
          type: ["string", "null"]
        somArtista:
          type: ["string", "null"]
        exclusivo:
          type: boolean
        totalCurtidas:
          type: integer
        totalComentarios:
          type: integer
        curtido:
          type: boolean
        guardado:
          type: boolean
        criadoEm:
          type: string
          format: date-time
    ReelInput:
      type: object
      required: [videoUrl]
      properties:
        videoUrl:
          type: string
        capaUrl:
          type: string
        legenda:
          type: string
        exclusivo:
          type: boolean
          default: false
    PaginatedReels:
      type: object
      required: [reels, page, hasMore]
      properties:
        reels:
          type: array
          items:
            $ref: "#/components/schemas/Reel"
        page:
          type: integer
        hasMore:
          type: boolean

    # SEARCH
...
    Highlight:
      type: object
      required: [id, titulo, criadoEm]
      properties:
        id:
          type: integer
        titulo:
          type: string
        capaUrl:
          type: ["string", "null"]
        stories:
          type: array
          items:
            $ref: "#/components/schemas/Story"
        criadoEm:
          type: string
          format: date-time
    HighlightInput:
      type: object
      required: [titulo, storyIds]
      properties:
        titulo:
          type: string
        capaUrl:
          type: string
        storyIds:
          type: array
          items:
            type: integer

    # CREATOR / MONETIZATION
...