# lib/api-zod/src/generated/types/storyView.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Xclusive API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSummary } from './userSummary';

export interface StoryView {
  utilizador: UserSummary;
  vistoEm: Date;
}